Revert to old ZipError to avoid breaking changes
